requirements.txt
albumentations==1.3.0
beautifulsoup4==4.11.1
bezier==2021.2.12
augmixations==0.1.2
editdistance==0.6.2
einops==0.6.0
filetype==1.2.0
lmdb==1.3.0
nltk==3.6.7
omegaconf==2.3.0
opencv-python==4.5.5.64
pillow
python-doctr==0.5.1
requests==2.27.1
scikit-learn==0.24.2
scikit-image==0.19.3
six==1.16.0
sympy==1.11.1
tabulate==0.9.0
torch==1.11.0
torchvision==0.12.0
transformers==4.26.1
pandas==1.4.2
wandb==0.13.10
warmup-scheduler==0.3
wget==3.2
